+++ shared/src/store/action-reducers/simulated-region.ts
@@ -19,6 +19,7 @@ import {
     PersonnelAvailableEvent,
     NewPatientEvent,
     MaterialAvailableEvent,
+    simulationBehaviorDictionary,
 } from '../../simulation';
 import { sendSimulationEvent } from '../../simulation/events/utils';
 import { cloneDeepMutable, UUID, uuidValidationOptions } from '../../utils';
@@ -337,6 +338,16 @@ export namespace SimulatedRegionActionReducers {
                         `The simulated region with id ${simulatedRegionId} has no behavior with id ${behaviorId}. Therefore it could not be removed.`
                     );
                 }
+
+                const behaviorState = simulatedRegion.behaviors[index]!;
+                if (simulationBehaviorDictionary[behaviorState.type].onRemove) {
+                    simulationBehaviorDictionary[behaviorState.type].onRemove!(
+                        draftState,
+                        simulatedRegion,
+                        behaviorState as any
+                    );
+                }
+
                 simulatedRegion.behaviors.splice(index, 1);
                 return draftState;
             },